Add corrected ErrorView with glass support
