+++ BrowserKit/Sources/SummarizeKit/SummarizeController.swift
@@ -78,78 +78,6 @@ public struct SummarizeViewModel {
     }
 }
 
-public struct ErrorViewModel {
-    let title: String
-    let titleA11yId: String
-    let actionButtonLabel: String
-    let actionButtonA11yId: String
-}
-
-class ErrorView: UIView,
-                 ThemeApplicable {
-    private struct UX {
-        static let labelHorizontalPadding: CGFloat = 44.0
-        static let actionButtonTopPadding: CGFloat = 32.0
-        static let actionButtonInsets = NSDirectionalEdgeInsets(
-            top: 12.0,
-            leading: 32.0,
-            bottom: 12.0,
-            trailing: 32.0
-        )
-    }
-    private let label: UILabel = .build {
-        $0.adjustsFontForContentSizeCategory = true
-        $0.font = FXFontStyles.Regular.body.scaledFont()
-        $0.numberOfLines = 0
-        $0.textAlignment = .center
-    }
-    private let actionButton: UIButton = .build {
-        $0.configuration = .filled()
-        $0.configuration?.cornerStyle = .capsule
-        $0.configuration?.contentInsets = UX.actionButtonInsets
-    }
-
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setup()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    private func setup() {
-        addSubviews(label, actionButton)
-
-        NSLayoutConstraint.activate([
-            label.topAnchor.constraint(equalTo: topAnchor),
-            label.leadingAnchor.constraint(equalTo: leadingAnchor, constant: UX.labelHorizontalPadding),
-            label.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -UX.labelHorizontalPadding),
-
-            actionButton.topAnchor.constraint(equalTo: label.bottomAnchor, constant: UX.actionButtonTopPadding),
-            actionButton.centerXAnchor.constraint(equalTo: centerXAnchor),
-            actionButton.leadingAnchor.constraint(greaterThanOrEqualTo: leadingAnchor),
-            actionButton.trailingAnchor.constraint(lessThanOrEqualTo: trailingAnchor),
-            actionButton.bottomAnchor.constraint(equalTo: bottomAnchor)
-        ])
-    }
-
-    func configure(viewModel: ErrorViewModel) {
-        label.text = viewModel.title
-        label.accessibilityIdentifier = viewModel.titleA11yId
-        actionButton.configuration?.title = viewModel.actionButtonLabel
-        actionButton.accessibilityIdentifier = viewModel.actionButtonA11yId
-    }
-
-    // MARK: - ThemeApplicable
-
-    func applyTheme(theme: any Theme) {
-        label.textColor = theme.colors.textOnDark
-        actionButton.configuration?.baseForegroundColor = theme.colors.textPrimary
-        actionButton.tintColor = theme.colors.actionTabActive
-    }
-}
-
 public class SummarizeController: UIViewController, Themeable {
     private struct UX {
         static let tabSnapshotInitialTransformPercentage: CGFloat = 0.5
@@ -180,13 +108,16 @@ public class SummarizeController: UIViewController, Themeable {
         $0.font = FXFontStyles.Regular.body.scaledFont()
         $0.alpha = 0
         $0.numberOfLines = 0
+        $0.textAlignment = .center
     }
     private let errorView: ErrorView = .build {
         $0.alpha = 0
     }
     private let closeButton: UIButton = .build {
+        if #available(iOS 26, *) {
+            $0.configuration = .prominentClearGlass()
+        }
         $0.adjustsImageSizeForAccessibilityContentSizeCategory = true
-        $0.alpha = 0
     }
     private let tabSnapshot: UIImageView = .build {
         $0.clipsToBounds = true
@@ -205,6 +136,11 @@ public class SummarizeController: UIViewController, Themeable {
             right: 0.0
         )
         $0.isEditable = false
+    }
+
+    // For the MVP only the portrait orientation is supported
+    override public var supportedInterfaceOrientations: UIInterfaceOrientationMask {
+        return .portrait
     }
 
     public init(
@@ -242,14 +178,6 @@ public class SummarizeController: UIViewController, Themeable {
     }
 
     private func setupSubviews() {
-        errorView.configure(
-            viewModel: .init(
-                title: "This is an error please reach oiyt to the team",
-                titleA11yId: "",
-                actionButtonLabel: "Continue",
-                actionButtonA11yId: ""
-            )
-        )
         view.addSubviews(tabSnapshot, gradient, closeButton, summaryView, loadingLabel, errorView)
 
         loadingLabel.text = viewModel.loadingLabel
@@ -287,22 +215,33 @@ public class SummarizeController: UIViewController, Themeable {
         tabSnapshotTopConstraint = tabSnapshot.topAnchor.constraint(equalTo: view.topAnchor)
         tabSnapshotTopConstraint?.isActive = true
 
+        let topHalfBoundGuide = UILayoutGuide()
+        view.addLayoutGuide(topHalfBoundGuide)
+
         NSLayoutConstraint.activate([
-            loadingLabel.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            loadingLabel.centerXAnchor.constraint(equalTo: view.centerXAnchor),
+            topHalfBoundGuide.topAnchor.constraint(equalTo: closeButton.bottomAnchor),
+            topHalfBoundGuide.bottomAnchor.constraint(equalTo: view.centerYAnchor),
+            topHalfBoundGuide.leadingAnchor.constraint(equalTo: view.leadingAnchor),
+            topHalfBoundGuide.trailingAnchor.constraint(equalTo: view.trailingAnchor),
+
+            loadingLabel.topAnchor.constraint(greaterThanOrEqualTo: topHalfBoundGuide.topAnchor),
             loadingLabel.leadingAnchor.constraint(greaterThanOrEqualTo: view.leadingAnchor,
                                                   constant: UX.summaryLabelHorizontalPadding),
+            loadingLabel.centerYAnchor.constraint(equalTo: topHalfBoundGuide.centerYAnchor),
+            loadingLabel.centerXAnchor.constraint(equalTo: topHalfBoundGuide.centerXAnchor),
             loadingLabel.trailingAnchor.constraint(lessThanOrEqualTo: view.trailingAnchor,
                                                    constant: -UX.summaryLabelHorizontalPadding),
+            loadingLabel.bottomAnchor.constraint(lessThanOrEqualTo: topHalfBoundGuide.bottomAnchor),
 
-            errorView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            errorView.centerXAnchor.constraint(equalTo: view.centerXAnchor),
-
+            errorView.centerXAnchor.constraint(equalTo: topHalfBoundGuide.centerXAnchor),
+            errorView.centerYAnchor.constraint(equalTo: topHalfBoundGuide.centerYAnchor),
+            errorView.topAnchor.constraint(greaterThanOrEqualTo: topHalfBoundGuide.topAnchor),
+            errorView.bottomAnchor.constraint(lessThanOrEqualTo: topHalfBoundGuide.bottomAnchor),
             errorView.leadingAnchor.constraint(greaterThanOrEqualTo: view.leadingAnchor),
             errorView.trailingAnchor.constraint(lessThanOrEqualTo: view.trailingAnchor),
 
-            closeButton.leadingAnchor.constraint(equalTo: view.leadingAnchor,
-                                                 constant: UX.closeButtonEdgePadding),
+            closeButton.trailingAnchor.constraint(equalTo: view.trailingAnchor,
+                                                  constant: -UX.closeButtonEdgePadding),
             closeButton.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor,
                                              constant: UX.closeButtonEdgePadding),
 
@@ -322,9 +261,7 @@ public class SummarizeController: UIViewController, Themeable {
         tabSnapshotTopConstraint?.constant = viewModel.tabSnapshotTopOffset
 
         let frameHeight = view.frame.height
-        loadingLabel.transform = CGAffineTransform(translationX: 0.0, y: frameHeight * 0.25)
         loadingLabel.startShimmering(light: .white, dark: .white.withAlphaComponent(0.1))
-        errorView.transform = CGAffineTransform(translationX: 0.0, y: frameHeight * 0.25)
 
         let transformAnimation = CABasicAnimation(keyPath: "transform.translation.y")
         transformAnimation.fromValue = 0
@@ -340,13 +277,12 @@ public class SummarizeController: UIViewController, Themeable {
 
         UIView.animate(withDuration: UX.initialTransformAnimationDuration, delay: 0.0, options: [], animations: {
             self.tabSnapshot.layer.cornerRadius = UX.tabSnapshotCornerRadius
-//            self.loadingLabel.alpha = 1.0
-            self.errorView.alpha = 1.0
+            self.loadingLabel.alpha = 1.0
         }) { _ in
             // TODO: - FXIOS-12858 replace this demo code with the actual backend API
-//            DispatchQueue.main.asyncAfter(deadline: .now() + 4.0) {
-//                self.showSummary()
-//            }
+            DispatchQueue.main.asyncAfter(deadline: .now() + 4.0) {
+                self.showSummary()
+            }
         }
     }
 
@@ -362,7 +298,6 @@ public class SummarizeController: UIViewController, Themeable {
            tabSnapshot.transform = CGAffineTransform(translationX: 0.0, y: tabSnapshotYTransform)
            loadingLabel.alpha = 0.0
            summaryView.alpha = 1.0
-           closeButton.alpha = 1.0
            view.backgroundColor = theme.colors.layer1
         } completion: { [weak self] _ in
             guard let tabSnapshotView = self?.tabSnapshot else { return }
